# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Indent Easy REST API
  version: 1.0.0
  summary: Authentication, files, webhooks, integrations and health endpoints of Indent Easy.
  description: |
    GraphQL (`/graphql`) is the primary client API; this document covers the REST surface only.

    Conventions
    - Base path `/api/v1`; JSON UTF-8; timestamps RFC 3339 UTC.
    - Errors use RFC 9457 `application/problem+json` extended with `code`, `requestId`, `details`.
    - `X-Request-ID` is accepted/echoed; `traceparent` (W3C) is propagated.
    - `Idempotency-Key` is required where the operation says so. Reusing a key with a different
      request body returns 409 `IDEMPOTENCY_KEY_REUSED`; with the same body the original response is replayed.
    - Rate limits are advertised with `RateLimit-Limit`, `RateLimit-Remaining`, `RateLimit-Reset`
      and `Retry-After` on 429.
  contact:
    name: Indent Easy Architecture Team
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:4000
    description: Local development
  - url: http://dev.indent-easy.local
    description: DEV (integration of the dev branch)
  - url: http://qa.indent-easy.local
    description: QA (testing and user acceptance)
  - url: http://prod.indent-easy.local
    description: PROD (interim host; HTTPS and the final domain arrive with the office server)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/files/upload-intents:
    post:
      tags: [Files]
      operationId: createUploadIntent
      summary: Create a document and obtain a pre-signed upload URL
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadIntentRequest'
      responses:
        '201':
          description: Upload intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadIntentResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          description: File exceeds the limit for the document type
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '415':
          description: MIME type not allowed for the document type
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/files/{documentId}/complete:
    post:
      tags: [Files]
      operationId: completeUpload
      summary: Confirm the object was uploaded; starts checksum verification and malware scan
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                multipartParts:
                  type: array
                  items:
                    type: object
                    required: [partNumber, etag]
                    properties:
                      partNumber:
                        type: integer
                      etag:
                        type: string
      responses:
        '202':
          description: Accepted; status PENDING_SCAN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentMeta'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          description: Uploaded object missing or checksum mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/files/{documentId}:
    get:
      tags: [Files]
      operationId: getDocument
      summary: Document metadata and status
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentMeta'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/files/{documentId}/download:
    get:
      tags: [Files]
      operationId: downloadDocument
      summary: Redirect to a short-lived signed download URL
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - name: version
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: disposition
          in: query
          required: false
          schema:
            type: string
            enum: [inline, attachment]
            default: attachment
      responses:
        '302':
          description: Redirect to signed URL (expires in 300 s)
          headers:
            Location:
              schema:
                type: string
                format: uri
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Document not AVAILABLE (pending scan or quarantined)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
    webhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        maxLength: 64
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
  responses:
    ValidationFailed:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.indent-easy/validation-failed
            title: Validation failed
            status: 400
            code: VALIDATION_FAILED
            detail: One or more fields are invalid.
            requestId: req-01926f6e8a1c
            details:
              fieldErrors:
                - path: [sizeBytes]
                  message: Must be less than or equal to 10485760
    Unauthorized:
      description: Authentication required or failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Not permitted
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Not found (also used when the caller may not know the resource exists)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Conflict (idempotency key reuse, version conflict, duplicate)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Stable machine-readable error code (same catalogue as GraphQL ErrorCode)
        requestId:
          type: string
        traceId:
          type: string
        details:
          type: object
          additionalProperties: true
...
    UploadIntentRequest:
      type: object
      required: [documentType, fileName, mimeType, sizeBytes, sha256]
      properties:
        documentType:
          type: string
          examples: [CHALLAN, INVOICE, POD, EWAY_BILL, APPROVAL_EVIDENCE, IMPORT_FILE]
        fileName:
          type: string
          maxLength: 255
        mimeType:
          type: string
        sizeBytes:
          type: integer
          minimum: 1
          maximum: 104857600
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
        entityType:
          type: string
        entityId:
          type: string
        existingDocumentId:
          type: string
          description: Upload a new version of an existing document
        multipart:
          type: boolean
          default: false
    UploadIntentResponse:
      type: object
      required: [documentId, versionNo, upload]
      properties:
        documentId:
          type: string
        versionNo:
          type: integer
        upload:
          type: object
          required: [method, expiresAt]
          properties:
            method:
              type: string
              enum: [PUT, MULTIPART]
            url:
              type: string
              format: uri
            headers:
              type: object
              additionalProperties:
                type: string
            parts:
              type: array
              items:
                type: object
                required: [partNumber, url]
                properties:
                  partNumber:
                    type: integer
                  url:
                    type: string
                    format: uri
            expiresAt:
              type: string
              format: date-time
    DocumentMeta:
      type: object
      required: [documentId, typeCode, fileName, mimeType, sizeBytes, status, versionNo]
      properties:
        documentId:
          type: string
        typeCode:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        sha256:
          type: string
        status:
          type: string
          enum: [PENDING_UPLOAD, PENDING_SCAN, AVAILABLE, QUARANTINED, SUPERSEDED, DELETED]
        versionNo:
          type: integer
        uploadedAt:
          type: string
          format: date-time
...